+++ src/main/java/org/seedstack/coffig/node/ValueNode.java
@@ -5,15 +5,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.coffig.node;
 
-import org.seedstack.coffig.TreeNode;
-import org.seedstack.coffig.internal.ConfigurationErrorCode;
-import org.seedstack.coffig.internal.ConfigurationException;
+package org.seedstack.coffig.node;
 
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Stream;
+import org.seedstack.coffig.TreeNode;
+import org.seedstack.coffig.internal.ConfigurationErrorCode;
+import org.seedstack.coffig.internal.ConfigurationException;
 
 public class ValueNode extends AbstractTreeNode {
     private final String value;
@@ -65,6 +65,10 @@ public class ValueNode extends AbstractTreeNode {
 
     @Override
     public Optional<TreeNode> get(String path) {
+        if (path.isEmpty()) {
+            return Optional.of(this);
+        }
+
         return Optional.empty();
     }